refactor: 인증코드 관련 패키지 및 클래스명 변경에 따른 수정
