+++ pennyway-app-external-api/src/main/java/kr/co/pennyway/api/apis/auth/service/AuthFindService.java
@@ -4,7 +4,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import kr.co.pennyway.api.apis.auth.dto.AuthFindDto;
-import kr.co.pennyway.domain.common.redis.phone.PhoneVerificationService;
 import kr.co.pennyway.domain.domains.user.domain.User;
 import kr.co.pennyway.domain.domains.user.exception.UserErrorCode;
 import kr.co.pennyway.domain.domains.user.exception.UserErrorException;
@@ -27,17 +26,6 @@ public class AuthFindService {
 	 */
 	@Transactional(readOnly = true)
 	public AuthFindDto.FindUsernameRes findUsername(String phone) {
-		// 인증번호 유효 여부 확인 > 인증번호 삭제
-		// try {
-		// 	if (phoneVerificationService.readByPhone(phone, PhoneVerificationType.FIND_USERNAME).equals(code)) {
-		// 		phoneVerificationService.delete(phone, PhoneVerificationType.FIND_USERNAME);
-		// 	}
-		// } catch (IllegalArgumentException e) {
-		// 	log.info("Phone verification code not found by phone: {}", phone);
-		// 	throw new PhoneVerificationException(PhoneVerificationErrorCode.EXPIRED_OR_INVALID_PHONE);
-		// }
-
-		// 사용자 조회
 		User user = userService.readUserByPhone(phone).orElseThrow(() -> {
 			log.info("User not found by phone: {}", phone);
 			return new UserErrorException(UserErrorCode.NOT_FOUND);
