+++ pennyway-app-external-api/src/main/java/kr/co/pennyway/api/apis/auth/usecase/AuthCheckUseCase.java
@@ -4,11 +4,11 @@ import org.springframework.transaction.annotation.Transactional;
 
 import kr.co.pennyway.api.apis.auth.dto.AuthFindDto;
 import kr.co.pennyway.api.apis.auth.dto.PhoneVerificationDto;
-import kr.co.pennyway.api.apis.auth.mapper.PhoneVerificationMapper;
 import kr.co.pennyway.api.apis.auth.service.AuthFindService;
+import kr.co.pennyway.api.apis.auth.service.PhoneVerificationService;
 import kr.co.pennyway.common.annotation.UseCase;
-import kr.co.pennyway.domain.common.redis.phone.PhoneVerificationService;
-import kr.co.pennyway.domain.common.redis.phone.PhoneVerificationType;
+import kr.co.pennyway.domain.common.redis.phone.PhoneCodeKeyType;
+import kr.co.pennyway.domain.common.redis.phone.PhoneCodeService;
 import kr.co.pennyway.domain.domains.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -19,8 +19,8 @@ import lombok.extern.slf4j.Slf4j;
 public class AuthCheckUseCase {
 	private final UserService userService;
 	private final AuthFindService authFindService;
-	private final PhoneVerificationMapper phoneVerificationMapper;
 	private final PhoneVerificationService phoneVerificationService;
+	private final PhoneCodeService phoneCodeService;
 
 	@Transactional(readOnly = true)
 	public boolean checkUsernameDuplicate(String username) {
@@ -29,8 +29,8 @@ public class AuthCheckUseCase {
 
 	@Transactional(readOnly = true)
 	public AuthFindDto.FindUsernameRes findUsername(String phone, String code) {
-		phoneVerificationMapper.isValidCode(PhoneVerificationDto.VerifyCodeReq.of(phone, code), PhoneVerificationType.FIND_USERNAME);
-		phoneVerificationService.delete(phone, PhoneVerificationType.FIND_USERNAME);
+		phoneVerificationService.isValidCode(PhoneVerificationDto.VerifyCodeReq.of(phone, code), PhoneCodeKeyType.FIND_USERNAME);
+		phoneCodeService.delete(phone, PhoneCodeKeyType.FIND_USERNAME);
 		return authFindService.findUsername(phone);
 	}
 }
